requirements.txt
pandas
sqlalchemy
python-binance
python-dotenv
asyncio
certifi
matplotlib